% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_patient_column.R
\name{add_patient_column}
\alias{add_patient_column}
\title{Add column from patient data}
\usage{
add_patient_column(df, df_patient, column_name)
}
\arguments{
\item{df}{A dataframe produced by clean_thermograms()}

\item{df_patient}{Patient data file generated by load_patient}

\item{column_name}{Column to be added}
}
\value{
Returns a dataframe
}
\description{
Add column from patient data
}
