% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_summary_column.R
\name{add_summary_column}
\alias{add_summary_column}
\title{Add column from summary data}
\usage{
add_summary_column(df, df_summary, column_name)
}
\arguments{
\item{df}{A dataframe produced by clean_thermograms()}

\item{df_summary}{Summary data file generated by generate_summary}

\item{column_name}{Column to be added}
}
\value{
Returns a dataframe
}
\description{
Add column from summary data
}
