% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_thermogram_lag.R
\name{add_thermogram_lag}
\alias{add_thermogram_lag}
\title{Add Thermogram Lag}
\usage{
add_thermogram_lag(df, column = "Lag", low_temp = "T45", high_temp = "T90")
}
\arguments{
\item{df}{A dataframe produced by clean_thermograms()}

\item{column}{The column name containing Lag data. Defaults to Lag.}

\item{low_temp}{The low temperature of the desired analysis. Defaults to 45.0}

\item{high_temp}{The high temperature of the desired analysis. Defaults to 90.0}
}
\value{
Returns a dataframe
}
\description{
Add Thermogram Lag
}
\examples{
add_thermogram_lag(df)
}
