% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_thermograms.R
\name{clean_thermograms}
\alias{clean_thermograms}
\title{Clean Thermograms}
\usage{
clean_thermograms(
  df,
  type = "Plasma",
  column = "SampleCode",
  low_temp = "T45",
  high_temp = "T90",
  temp_range = seq(45, 90, by = 0.1),
  summary = c("tarea", "fwhm", "max", "tpeakf", "peakf", "tpeak1", "peak1", "tpeak2",
    "peak2", "tpeak3", "peak3", "P1P2.trough", "tP1P2.trough", "tmax", "tfm",
    "peak12ratio", "min", "tmin", "median"),
  lag = FALSE
)
}
\arguments{
\item{df}{A dataframe produced by load_thermograms()}

\item{type}{The type of thermogram. Defaults to Plasma}

\item{column}{The column name containing patient/thermogram identifier. Defaults to SampleCode.}

\item{low_temp}{The low temperature of the desired analysis. Defaults to 45.0}

\item{high_temp}{The high temperature of the desired analysis. Defaults to 90.0}

\item{temp_range}{The range of temperature and step}

\item{summary}{A vector containing desired summary statistics, defaults to NULL}

\item{lag}{A TRUE or FALSE to include lag information, defaults to FALSE}
}
\value{
Returns a dataframe
}
\description{
Clean Thermograms
}
