% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_peak.R
\name{gen_peak}
\alias{gen_peak}
\title{Generate Peak}
\usage{
gen_peak(x, l.temp, h.temp)
}
\arguments{
\item{x}{Value}

\item{l.temp}{Low temperature range}

\item{h.temp}{High temperature range}
}
\value{
Returns a peak
}
\description{
Generate Peak
}
