% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_summary.R
\name{generate_summary}
\alias{generate_summary}
\title{Generate Summary}
\usage{
generate_summary(df, type, sample_code, temperatures)
}
\arguments{
\item{df}{A dataframe produced by load_thermograms()}

\item{type}{The type of thermogram. Defaults to Plasma}

\item{sample_code}{TBD}

\item{temperatures}{The range of temperatures for the thermograms. Defaults to 45.0 - 90.0 by 0.1}
}
\value{
Returns a dataframe
}
\description{
Generate Summary
}
