% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_patientdata.R
\name{load_patientdata}
\alias{load_patientdata}
\title{Load Patient Data}
\usage{
load_patientdata(file, sheet = "Sheet1")
}
\arguments{
\item{file}{The path and file name of a correctly formatted .xlsx or.csv file}

\item{sheet}{The Sheet Name in the Excel file. Defaults to Sheet1}
}
\value{
Returns a dataframe
}
\description{
Load Patient Data
}
\examples{
load_patientdata(file='../GenSumv21/data-raw/patient_urine_y4.xlsx')
}
