% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_thermograms.R
\name{load_thermograms}
\alias{load_thermograms}
\title{Load Thermograms into a Dataframe}
\usage{
load_thermograms(
  file,
  sheet = "Sheet1",
  blank_row = 1,
  temps = c(paste0("T", seq(21, 110, by = 0.1)))
)
}
\arguments{
\item{file}{The path and file name of a correctly formatted .xlsx or.csv file}

\item{sheet}{The Sheet Name in the Excel file. Defaults to Sheet1}

\item{blank_row}{Number of blank rows due to transpose. Defaults to 1.}

\item{temps}{The range of temperatures for the thermograms. Defaults to 21.0 - 110.0 by 0.1}
}
\value{
Returns a dataframe
}
\description{
Load Thermograms into a Dataframe
}
\examples{
load_thermograms(file='../GenSumv21/data-raw/plasma_thermo_y3.xlsx')
(c(paste0('T', seq(45.0, 90.0, by=0.1)), 'binary'))
}
